ranking a partir de uma view

+++ suggestions-web/src/main/java/ifpb/pos/suggestions/models/RankedUser.java
@@ -8,24 +8,34 @@ package ifpb.pos.suggestions.models;
 import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.NamedQueries;
+import javax.persistence.NamedQuery;
+import javax.persistence.Table;
 
 /**
  *
  * @author natarajan
  */
 @Entity
+//@Table(name = "ranking")
+//@NamedQuery(name = "Ranking.all", query = "FROM Ranking")
+//@NamedQueries({
+//    @NamedQuery(name = "Ranking.all", query = "FROM ranking")
+//        , 
+//    @NamedQuery(name = "Ranking.user", query = "FROM ranking WHERE iduser = :userapp")
+//})
 public class RankedUser implements Serializable{
     private static final long serialVersionUID = 1989L;
     
     @Id
     private Long idUser;
     private double ranking;
-    private int rankingPosition;
+    private Long rankingPosition;
 
     public RankedUser() {
     }
 
-    public RankedUser(Long idUser, double ranking, int rankingPosition) {
+    public RankedUser(Long idUser, double ranking, Long rankingPosition) {
         this.idUser = idUser;
         this.ranking = ranking;
         this.rankingPosition = rankingPosition;
@@ -47,11 +57,11 @@ public class RankedUser implements Serializable{
         this.ranking = ranking;
     }
 
-    public int getRankingPosition() {
+    public Long getRankingPosition() {
         return rankingPosition;
     }
 
-    public void setRankingPosition(int rankingPosition) {
+    public void setRankingPosition(Long rankingPosition) {
         this.rankingPosition = rankingPosition;
     }
